module: add custom runtime blacklist logging for ZRAM and ZSMALLOC
Introduce a custom hardcoded blacklist in the module loader that
matches "zram" and "zsmalloc". When these modules are requested,
the loader prints a log message indicating that the module is
blacklisted, but still allows it to be loaded since the failure
return (-EPERM) is disabled.

Signed-off by: Rohail33 <rohailyousafzai49@gmail.com>